Moved iterators into the correct files to maintain proper codebase visibility
