+++ networkx/algorithms/__init__.py
@@ -125,4 +125,3 @@ from networkx.algorithms.tree.decomposition import *
 from networkx.algorithms.tree.mst import *
 from networkx.algorithms.tree.operations import *
 from networkx.algorithms.tree.recognition import *
-from networkx.algorithms.tree.tree_iterators import *
