+++ networkx/algorithms/tree/__init__.py
@@ -1,7 +1,6 @@
 from .branchings import *
 from .coding import *
 from .mst import *
-from .tree_iterators import *
 from .recognition import *
 from .operations import *
 from .decomposition import *
